chore(scripts/dbdeploy): separate db-audit from db-sample
