+++ scripts/dbdeploy/db-audit/.gitignore
@@ -1,0 +1,2 @@
+dev.db
+.env
