+++ scripts/dbdeploy/db-sample/drizzle/meta/_journal.json
@@ -22,6 +22,13 @@
       "when": 1747612800000,
       "tag": "0002_remove_rbac_fga",
       "breakpoints": true
+    },
+    {
+      "idx": 3,
+      "version": "7",
+      "when": 1747699200000,
+      "tag": "0003_remove_audit",
+      "breakpoints": true
     }
   ]
 }
