+++ scripts/dbdeploy/package.json
@@ -15,6 +15,10 @@
     "db:iam:migrate": "cd db-iam && drizzle-kit migrate",
     "db:iam:studio": "cd db-iam && drizzle-kit studio",
     "db:iam:seed": "node --env-file=db-iam/.env db-iam/seed.ts",
+    "db:audit:generate": "cd db-audit && drizzle-kit generate",
+    "db:audit:migrate": "cd db-audit && drizzle-kit migrate",
+    "db:audit:studio": "cd db-audit && drizzle-kit studio",
+    "db:audit:seed": "node --env-file=db-audit/.env db-audit/seed.ts",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
